+++ packages/archives/CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	* `@otuekong-portfolio/infrastructure-server`
 
 ### Changed
+* Applying 'Chip' to 'Pill' changes
+* Renaming `TechnologyChips` component to `TechnologyPills`
 * Updating `BioSection` component to display both bio text and `children`, when provided
 * Updating bio text to abstract metadata
 * Updating author information in `package.json`